NEB service relate functions moved to qcf_neb.py. ComputeMetric deleted.
